Updated UIManager.java
Added taskDelay and taskPeriod to openGUI(Player, UserInterface), to enable customisation of the Update task without having to refactor code.

Removed UserInterface reference on getItem(UIEntry).

Added Exceptions to setupInventory which better enables Developers to handle errors.

Added isValidInventorySize(int) method to test if an integer is a valid Bukkit compatible inventory size.

Added 1.7.X and 1.8.X features to the API to enable Coloured Wool for older versions, etc. etc.

Added null check for the ItemMeta within getItem(UIEntry) because it can return Null on broken occasions.

Added isExperimental() which determines if the server is running 1.7.X or 1.8.X server software.

Changed cancelTask slightly as there is no check if the UUID actually has a task which could cancel tasks which could hurt other systems on the server.